Move metadata repository and types under metadata folder
